\name{dat.lim2014}
\docType{data}
\alias{dat.lim2014.1}
\alias{dat.lim2014.2}
\alias{dat.lim2014.3}
\alias{dat.lim2014.4}
\alias{dat.lim2014.phy1}
\alias{dat.lim2014.phy2}
\alias{dat.lim2014.phy3}
\alias{dat.lim2014.phy4}
\alias{dat.lim2014}
\title{Heterogeneity in individual quality and reproductive trade‐offs within species}
\description{Correlations between: 1) maternal size and offspring size (dat.lim2014.1), 2) maternal size and maternal fecundity (dat.lim2014.2), 3) offspring size and offspring number (dat.lim2014.3), 4) offspring size and offspring number, adjusted for maternal size (dat.lim2014.4)}
\usage{dat.lim2014}
\format{
The data frame dat.lim2014.1 contains the following columns:
\tabular{lll}{
\bold{Article} \tab \code{integer} \tab a numeric identifier denoting the primary article \cr
\bold{Datapoint} \tab \code{integer} \tab a unqiue numeric identifier for the specific correlation \cr
\bold{Amniotes} \tab \code{character} \tab whether the species is amniotic or not \cr
\bold{animal} \tab \code{character} \tab the species in which the correlation is made \cr
\bold{Common.Name} \tab \code{character} \tab ADD_DESCRIPTION \cr
\bold{N} \tab \code{integer} \tab ADD_DESCRIPTION \cr
\bold{Maternal} \tab \code{character} \tab ADD_DESCRIPTION \cr
\bold{Offspring} \tab \code{character} \tab ADD_DESCRIPTION \cr
\bold{Reproduction} \tab \code{character} \tab ADD_DESCRIPTION \cr
\bold{Propagule} \tab \code{character} \tab ADD_DESCRIPTION \cr
\bold{Environment} \tab \code{character} \tab ADD_DESCRIPTION \cr
\bold{Author} \tab \code{character} \tab ADD_DESCRIPTION \cr
\bold{year} \tab \code{character} \tab ADD_DESCRIPTION \cr
\bold{Phylum} \tab \code{character} \tab ADD_DESCRIPTION \cr
\bold{Class} \tab \code{character} \tab ADD_DESCRIPTION \cr
\bold{Order} \tab \code{character} \tab ADD_DESCRIPTION \cr
\bold{Family} \tab \code{character} \tab ADD_DESCRIPTION \cr
\bold{Genus} \tab \code{character} \tab ADD_DESCRIPTION \cr
\bold{RU} \tab \code{character} \tab ADD_DESCRIPTION \cr
\bold{yi} \tab \code{numeric} \tab ADD_DESCRIPTION \cr
}  
The data frame dat.lim2014.2 contains the following columns:
\tabular{lll}{
\bold{Article} \tab \code{integer} \tab ADD_DESCRIPTION \cr
\bold{Datapoint} \tab \code{integer} \tab ADD_DESCRIPTION \cr
\bold{Amniotes} \tab \code{character} \tab ADD_DESCRIPTION \cr
\bold{animal} \tab \code{character} \tab ADD_DESCRIPTION \cr
\bold{Common.Name} \tab \code{character} \tab ADD_DESCRIPTION \cr
\bold{N} \tab \code{integer} \tab ADD_DESCRIPTION \cr
\bold{Maternal} \tab \code{character} \tab ADD_DESCRIPTION \cr
\bold{Fecundity} \tab \code{character} \tab ADD_DESCRIPTION \cr
\bold{Reproduction} \tab \code{character} \tab ADD_DESCRIPTION \cr
\bold{Propagule} \tab \code{character} \tab ADD_DESCRIPTION \cr
\bold{Environment} \tab \code{character} \tab ADD_DESCRIPTION \cr
\bold{Author} \tab \code{character} \tab ADD_DESCRIPTION \cr
\bold{year} \tab \code{character} \tab ADD_DESCRIPTION \cr
\bold{Phylum} \tab \code{character} \tab ADD_DESCRIPTION \cr
\bold{Class} \tab \code{character} \tab ADD_DESCRIPTION \cr
\bold{Order} \tab \code{character} \tab ADD_DESCRIPTION \cr
\bold{Family} \tab \code{character} \tab ADD_DESCRIPTION \cr
\bold{Genus} \tab \code{character} \tab ADD_DESCRIPTION \cr
\bold{RU} \tab \code{character} \tab ADD_DESCRIPTION \cr
\bold{yi} \tab \code{numeric} \tab ADD_DESCRIPTION \cr
}  
The data frame dat.lim2014.3 contains the following columns:
\tabular{lll}{
\bold{Article} \tab \code{integer} \tab ADD_DESCRIPTION \cr
\bold{Amniotes} \tab \code{character} \tab ADD_DESCRIPTION \cr
\bold{animal} \tab \code{character} \tab ADD_DESCRIPTION \cr
\bold{N} \tab \code{integer} \tab ADD_DESCRIPTION \cr
\bold{Maternal} \tab \code{character} \tab ADD_DESCRIPTION \cr
\bold{Offspring} \tab \code{character} \tab ADD_DESCRIPTION \cr
\bold{Propagule} \tab \code{character} \tab ADD_DESCRIPTION \cr
\bold{Reproduction} \tab \code{character} \tab ADD_DESCRIPTION \cr
\bold{Environment} \tab \code{character} \tab ADD_DESCRIPTION \cr
\bold{Author} \tab \code{character} \tab ADD_DESCRIPTION \cr
\bold{Date} \tab \code{character} \tab ADD_DESCRIPTION \cr
\bold{Species} \tab \code{character} \tab ADD_DESCRIPTION \cr
\bold{Phylum} \tab \code{character} \tab ADD_DESCRIPTION \cr
\bold{Class} \tab \code{character} \tab ADD_DESCRIPTION \cr
\bold{Order} \tab \code{character} \tab ADD_DESCRIPTION \cr
\bold{Family} \tab \code{character} \tab ADD_DESCRIPTION \cr
\bold{Genus} \tab \code{character} \tab ADD_DESCRIPTION \cr
\bold{yi} \tab \code{numeric} \tab ADD_DESCRIPTION \cr
}  
The data frame dat.lim2014.4 contains the following columns:
\tabular{lll}{
\bold{Article} \tab \code{integer} \tab ADD_DESCRIPTION \cr
\bold{Amniotes} \tab \code{character} \tab ADD_DESCRIPTION \cr
\bold{animal} \tab \code{character} \tab ADD_DESCRIPTION \cr
\bold{N} \tab \code{integer} \tab ADD_DESCRIPTION \cr
\bold{Maternal} \tab \code{character} \tab ADD_DESCRIPTION \cr
\bold{Offspring} \tab \code{character} \tab ADD_DESCRIPTION \cr
\bold{Propagule} \tab \code{character} \tab ADD_DESCRIPTION \cr
\bold{Reproduction} \tab \code{character} \tab ADD_DESCRIPTION \cr
\bold{Environment} \tab \code{character} \tab ADD_DESCRIPTION \cr
\bold{Author} \tab \code{character} \tab ADD_DESCRIPTION \cr
\bold{Date} \tab \code{character} \tab ADD_DESCRIPTION \cr
\bold{Species} \tab \code{character} \tab ADD_DESCRIPTION \cr
\bold{Phylum} \tab \code{character} \tab ADD_DESCRIPTION \cr
\bold{Class} \tab \code{character} \tab ADD_DESCRIPTION \cr
\bold{Order} \tab \code{character} \tab ADD_DESCRIPTION \cr
\bold{Family} \tab \code{character} \tab ADD_DESCRIPTION \cr
\bold{Genus} \tab \code{character} \tab ADD_DESCRIPTION \cr
\bold{yi} \tab \code{numeric} \tab ADD_DESCRIPTION \cr
}  
The data dat.lim2014.phy1 contains a phylogeny for the species in dat.lim2014.1\cr
The data dat.lim2014.phy2 contains a phylogeny for the species in dat.lim2014.2\cr
The data dat.lim2014.phy3 contains a phylogeny for the species in dat.lim2014.3\cr
The data dat.lim2014.phy4 contains a phylogeny for the species in dat.lim2014.4\cr
}
\details{ADD_DETAILS}
\source{Lim, J.N., Senior, A.M. & Nakagawa, S. (2014). Heterogeneity in individual quality and reproductive trade‐offs within species. \emph{Evolution}, \bold{8}, 2306-2318.}
\author{Alistair M Senior, \email{alistair.senior@sydney.edu.au}}
\examples{
### copy data into 'dat' and examine data
dat <- dat.lim2014
dat
}
\keyword{datasets}
\concept{ecology}
\concept{evolution}
\concept{phylogeny}
\concept{correlation}
