\name{dat.maire2019}
\docType{data}
\alias{dat.maire2019.data}
\alias{dat.maire2019.dmat}
\alias{dat.maire2019}
\title{ADD_TITLE}
\description{ADD_DESCRIPTION}
\usage{dat.maire2019}
\format{
The data dat.maire2019.data contains ADD_DETAILS\cr
The data dat.maire2019.dmat contains ADD_DETAILS\cr
}
\details{ADD_DETAILS}
\source{ADD_REFERENCE}
\author{ADD_CONTRIBUTOR_NAME, \email{ADD_EMAIL}}
\examples{
### copy data into 'dat' and examine data
dat <- dat.maire2019.data
dat
}
\keyword{datasets}
\concept{ADD_CONCEPT}
