\name{dat.sorokowski2019}
\docType{data}
\alias{dat.sorokowski2019}
\title{Studies on Sex Differences in Human Olfaction}
\description{Sex Differences in Human Olfaction including discrimination, threshold, and identification.}
\usage{dat.sorokowski2019}
\format{
The data frame dat.sorokowski2019.1 contains data for olfactory discrimination and consists of the following columns:
\tabular{lll}{
\bold{ID}        \tab \code{numeric}   \tab Study ID number \cr
\bold{Author}     \tab \code{character} \tab Author or authors of the paper \cr
\bold{Year}        \tab \code{numeric}   \tab Year of publication \cr
\bold{Paper Title}        \tab \code{character}   \tab Title of paper \cr
\bold{Country}        \tab \code{character}   \tab Country where study was conducted \cr
\bold{Nwomen}        \tab \code{numeric}   \tab Number of women \cr
\bold{Nmen}        \tab \code{numeric}   \tab Number of men \cr
\bold{meanWomen}        \tab \code{numeric}   \tab Mean score for women \cr
\bold{sdWomen}        \tab \code{numeric}   \tab Standard deviation for women \cr
\bold{ageMeanWomen}        \tab \code{numeric}   \tab Mean age for women \cr
\bold{ageSDWomen}        \tab \code{numeric}   \tab Standard deviation of age for women \cr
\bold{meanMen}        \tab \code{numeric}   \tab Mean score for men \cr
\bold{sdMen}        \tab \code{numeric}   \tab Standard deviation for men \cr
\bold{ageMeanMen}        \tab \code{numeric}   \tab Mean age for men \cr
\bold{ageSDMen}        \tab \code{numeric}   \tab Standard deviation of age for men \cr


}
}
\details{
   The three datasets included in this entry are used to determin differences in human olfaction based on sex including identification, discrimination, and threshold. 
   
   Data and Jamovi analysis files can be found here https://osf.io/6tfuy.
}
\source{
  Sorokowski, P., Karwowski, M., Misiak, M., Marczak, M. K., Dziekan, M., Hummel, T., & Sorokowska, A. (2019). Sex differences in human olfaction: A meta-analysis. \emph{Frontiers in Psychology}, 10, 242.
}
\keyword{datasets}
