\name{dat_search}
\alias{dat_search}
\title{Searching data function}
\description{Searches all existing data files based on concept descriptors. Prompts users to }
\usage{
dat_search("ecology")
}
\arguments{
   \item{pattern}{Character string specifying the search pattern to match within the data files using fuzzy matching.}
   \item{fields}{The specific field or fields in the documentation to search. The default is the "concept" field.}
}
\details{
   The function ...
}
\keyword{file}