\name{prep_data}
\alias{prep_data}
\title{Data preparation function}
\description{Function to run data processing scripts.}
\usage{
prep_data(rebuild=FALSE, overwrite=NULL)
}
\arguments{
   \item{rebuild}{logical indicating whether the entire database should be rebuild (default is \code{FALSE}.}
   \item{overwrite}{character vector with one or more \code{.Rd} filenames to overwrite (if they already exist).}
}
\details{
   The function ...
}
\keyword{file}
