\name{url_dl}
\alias{url_dl}
\title{URL Delivered Documents}
\usage{
  url_dl(..., url = "http://dl.dropbox.com/u/61803503/")
}
\arguments{
  \item{\ldots}{Files for download.}

  \item{url}{The url to pass information from.}
}
\value{
  Downloads a file to the users working directory.
}
\description{
  A convenience function for passing new data sets to users
  in the future.
}

