\name{dat.lim2014}
\docType{data}
\alias{dat.lim2014.1}
\alias{dat.lim2014.2}
\alias{dat.lim2014.3}
\alias{dat.lim2014.4}
\alias{dat.lim2014.phy1}
\alias{dat.lim2014.phy2}
\alias{dat.lim2014.phy3}
\alias{dat.lim2014.phy4}
\alias{dat.lim2014}
\title{Heterogeneity in individual quality and reproductive trade‐offs within species}
\description{Correlations between: 1) maternal size and offspring size (dat.lim2014.1), 2) maternal size and maternal fecundity (dat.lim2014.2), 3) offspring size and offspring number (dat.lim2014.3), 4) offspring size and offspring number, adjusted for maternal size (dat.lim2014.4)}
\usage{dat.lim2014}
\format{
The data frame dat.lim2014.1 contains the following columns:
\tabular{lll}{
\bold{Article} \tab \code{integer} \tab a numeric identifier denoting the primary article \cr
\bold{Datapoint} \tab \code{integer} \tab a unqiue numeric identifier for the specific correlation \cr
\bold{Amniotes} \tab \code{character} \tab whether the species is amniotic or not \cr
\bold{animal} \tab \code{character} \tab the species in which the correlation is made \cr
\bold{Common.Name} \tab \code{character} \tab common name of the species \cr
\bold{N} \tab \code{integer} \tab the sample size of the correlation \cr
\bold{Maternal} \tab \code{character} \tab the trait used to quantify maternal size \cr
\bold{Offspring} \tab \code{character} \tab the trait used to quantify offspring size \cr
\bold{Reproduction} \tab \code{character} \tab does the species reproduce semalparously or iteroparously \cr
\bold{Propagule} \tab \code{character} \tab the life stage of the offspring \cr
\bold{Environment} \tab \code{character} \tab is the correlaiton quanitifed in the wild or in captivity \cr
\bold{Author} \tab \code{character} \tab authors of the primary study \cr
\bold{year} \tab \code{character} \tab year published \cr
\bold{Phylum} \tab \code{character} \tab phylum of the species \cr
\bold{Class} \tab \code{character} \tab class of the species \cr
\bold{Order} \tab \code{character} \tab order of the species \cr
\bold{Family} \tab \code{character} \tab family of the species \cr
\bold{Genus} \tab \code{character} \tab genus of the species \cr
\bold{RU} \tab \code{character} \tab single of lifetime reproductive unit \cr
\bold{yi} \tab \code{numeric} \tab effect size of correlation (zr) \cr
}  
The data frame dat.lim2014.2 contains the following columns:
\tabular{lll}{
\bold{Article} \tab \code{integer} \tab a numeric identifier denoting the primary article \cr
\bold{Datapoint} \tab \code{integer} \tab a unqiue numeric identifier for the specific correlation \cr
\bold{Amniotes} \tab \code{character} \tab whether the species is amniotic or not \cr
\bold{animal} \tab \code{character} \tab the species in which the correlation is made \cr
\bold{Common.Name} \tab \code{character} \tab common name of the species \cr
\bold{N} \tab \code{integer} \tab the sample size of the correlation \cr
\bold{Maternal} \tab \code{character} \tab the trait used to quantify maternal size \cr
\bold{Fecundity} \tab \code{character} \tab the trait used to quantify offspring numbers \cr
\bold{Reproduction} \tab \code{character} \tab does the species reproduce semalparously or iteroparously \cr
\bold{Propagule} \tab \code{character} \tab the life stage of the offspring \cr
\bold{Environment} \tab \code{character} \tab is the correlaiton quanitifed in the wild or in captivity \cr
\bold{Author} \tab \code{character} \tab authors of the primary study \cr
\bold{year} \tab \code{character} \tab year published \cr
\bold{Phylum} \tab \code{character} \tab phylum of the species \cr
\bold{Class} \tab \code{character} \tab class of the species \cr
\bold{Order} \tab \code{character} \tab order of the species \cr
\bold{Family} \tab \code{character} \tab family of the species \cr
\bold{Genus} \tab \code{character} \tab genus of the species \cr
\bold{RU} \tab \code{character} \tab single of lifetime reproductive unit \cr
\bold{yi} \tab \code{numeric} \tab effect size of correlation (zr) \cr
}  
The data frame dat.lim2014.3 contains the following columns:
\tabular{lll}{
\bold{Article} \tab \code{integer} \tab a numeric identifier denoting the primary article \cr
\bold{Amniotes} \tab \code{character} \tab whether the species is amniotic or not \cr
\bold{animal} \tab \code{character} \tab the species in which the correlation is made \cr
\bold{N} \tab \code{integer} \tab the sample size of the correlation \cr
\bold{Maternal} \tab \code{character} \tab the trait used to quantify maternal size \cr
\bold{Offspring} \tab \code{character} \tab the trait used to quantify offspring size \cr
\bold{Propagule} \tab \code{character} \tab the life stage of the offspring \cr
\bold{Reproduction} \tab \code{character} \tab does the species reproduce semalparously or iteroparously \cr
\bold{Environment} \tab \code{character} \tab is the correlaiton quanitifed in the wild or in captivity \cr
\bold{Author} \tab \code{character} \tab authors of the primary study \cr
\bold{Date} \tab \code{character} \tab year published \cr
\bold{Phylum} \tab \code{character} \tab phylum of the species \cr
\bold{Class} \tab \code{character} \tab class of the species \cr
\bold{Order} \tab \code{character} \tab order of the species \cr
\bold{Family} \tab \code{character} \tab family of the species \cr
\bold{Genus} \tab \code{character} \tab genus of the species \cr
\bold{yi} \tab \code{numeric} \tab effect size of correlation (zr) \cr
}  
The data frame dat.lim2014.4 contains the following columns:
\tabular{lll}{
\bold{Article} \tab \code{integer} \tab a numeric identifier denoting the primary article \cr
\bold{Amniotes} \tab \code{character} \tab whether the species is amniotic or not \cr
\bold{animal} \tab \code{character} \tab the species in which the correlation is made \cr
\bold{N} \tab \code{integer} \tab the sample size of the correlation \cr
\bold{Maternal} \tab \code{character} \tab the trait used to quantify maternal size \cr
\bold{Offspring} \tab \code{character} \tab the trait used to quantify offspring size \cr
\bold{Propagule} \tab \code{character} \tab the life stage of the offspring \cr
\bold{Reproduction} \tab \code{character} \tab does the species reproduce semalparously or iteroparously \cr
\bold{Environment} \tab \code{character} \tab is the correlaiton quanitifed in the wild or in captivity \cr
\bold{Author} \tab \code{character} \tab authors of the primary study \cr
\bold{Date} \tab \code{character} \tab year published \cr
\bold{Phylum} \tab \code{character} \tab phylum of the species \cr
\bold{Class} \tab \code{character} \tab class of the species \cr
\bold{Order} \tab \code{character} \tab order of the species \cr
\bold{Family} \tab \code{character} \tab family of the species \cr
\bold{Genus} \tab \code{character} \tab genus of the species \cr
\bold{yi} \tab \code{numeric} \tab effect size of correlation (zr) \cr
}  
The data dat.lim2014.phy1 contains a phylogeny for the species in dat.lim2014.1\cr
The data dat.lim2014.phy2 contains a phylogeny for the species in dat.lim2014.2\cr
The data dat.lim2014.phy3 contains a phylogeny for the species in dat.lim2014.3\cr
The data dat.lim2014.phy4 contains a phylogeny for the species in dat.lim2014.4\cr
}
\details{
	A meta-analysis on the strength of correlation between 1) maternal size and offspring size (dat.lim2014.1), 2) maternal size and maternal fecundity (dat.lim2014.2), 3) offspring size and offspring number (dat.lim2014.3), 4) offspring size and offspring number, adjusted for maternal size (dat.lim2014.4).
	
	In each case the correlation is given as the z-transformed correlation coefficient. In places these have been computed from other statsitics in the primary literature. A number of predictors related to life-history traits of the species are given. Topologies for the phylogenetic relationship between the species included in each data set are included. 
	}
\source{Lim, J.N., Senior, A.M. & Nakagawa, S. (2014). Heterogeneity in individual quality and reproductive trade‐offs within species. \emph{Evolution}, \bold{8}, 2306-2318.}
\author{Alistair M Senior, \email{alistair.senior@sydney.edu.au}}
\examples{
### copy data into 'dat' and examine data
dat <- dat.lim2014
dat
}
\keyword{datasets}
\concept{ecology}
\concept{evolution}
\concept{phylogeny}
\concept{correlation}
