\name{dat.lim2014}
\docType{data}
\alias{dat.lim2014.1}
\alias{dat.lim2014.2}
\alias{dat.lim2014.3}
\alias{dat.lim2014.4}
\alias{dat.lim2014.phy1}
\alias{dat.lim2014.phy2}
\alias{dat.lim2014.phy3}
\alias{dat.lim2014.phy4}
\alias{dat.lim2014}
\title{ADD_TITLE}
\description{ADD_DESCRIPTION}
\usage{dat.lim2014}
\format{
The data frame dat.lim2014.1 contains the following columns:
\tabular{lll}{
\bold{Article} \tab \code{integer} \tab ADD_DESCRIPTION \cr
\bold{Datapoint} \tab \code{integer} \tab ADD_DESCRIPTION \cr
\bold{Amniotes} \tab \code{character} \tab ADD_DESCRIPTION \cr
\bold{animal} \tab \code{character} \tab ADD_DESCRIPTION \cr
\bold{Species} \tab \code{character} \tab ADD_DESCRIPTION \cr
\bold{Common.Name} \tab \code{character} \tab ADD_DESCRIPTION \cr
\bold{N} \tab \code{integer} \tab ADD_DESCRIPTION \cr
\bold{Maternal} \tab \code{character} \tab ADD_DESCRIPTION \cr
\bold{Offspring} \tab \code{character} \tab ADD_DESCRIPTION \cr
\bold{Reproduction} \tab \code{character} \tab ADD_DESCRIPTION \cr
\bold{Propagule} \tab \code{character} \tab ADD_DESCRIPTION \cr
\bold{Environment} \tab \code{character} \tab ADD_DESCRIPTION \cr
\bold{Author} \tab \code{character} \tab ADD_DESCRIPTION \cr
\bold{year} \tab \code{character} \tab ADD_DESCRIPTION \cr
\bold{Phylum} \tab \code{character} \tab ADD_DESCRIPTION \cr
\bold{Class} \tab \code{character} \tab ADD_DESCRIPTION \cr
\bold{Order} \tab \code{character} \tab ADD_DESCRIPTION \cr
\bold{Family} \tab \code{character} \tab ADD_DESCRIPTION \cr
\bold{Genus} \tab \code{character} \tab ADD_DESCRIPTION \cr
\bold{RU} \tab \code{character} \tab ADD_DESCRIPTION \cr
\bold{yi} \tab \code{numeric} \tab ADD_DESCRIPTION \cr
}  
The data frame dat.lim2014.2 contains the following columns:
\tabular{lll}{
\bold{Article} \tab \code{integer} \tab ADD_DESCRIPTION \cr
\bold{Datapoint} \tab \code{integer} \tab ADD_DESCRIPTION \cr
\bold{Amniotes} \tab \code{character} \tab ADD_DESCRIPTION \cr
\bold{animal} \tab \code{character} \tab ADD_DESCRIPTION \cr
\bold{Species} \tab \code{character} \tab ADD_DESCRIPTION \cr
\bold{Common.Name} \tab \code{character} \tab ADD_DESCRIPTION \cr
\bold{N} \tab \code{integer} \tab ADD_DESCRIPTION \cr
\bold{Maternal} \tab \code{character} \tab ADD_DESCRIPTION \cr
\bold{Fecundity} \tab \code{character} \tab ADD_DESCRIPTION \cr
\bold{Reproduction} \tab \code{character} \tab ADD_DESCRIPTION \cr
\bold{Propagule} \tab \code{character} \tab ADD_DESCRIPTION \cr
\bold{Environment} \tab \code{character} \tab ADD_DESCRIPTION \cr
\bold{Author} \tab \code{character} \tab ADD_DESCRIPTION \cr
\bold{year} \tab \code{character} \tab ADD_DESCRIPTION \cr
\bold{Phylum} \tab \code{character} \tab ADD_DESCRIPTION \cr
\bold{Class} \tab \code{character} \tab ADD_DESCRIPTION \cr
\bold{Order} \tab \code{character} \tab ADD_DESCRIPTION \cr
\bold{Family} \tab \code{character} \tab ADD_DESCRIPTION \cr
\bold{Genus} \tab \code{character} \tab ADD_DESCRIPTION \cr
\bold{RU} \tab \code{character} \tab ADD_DESCRIPTION \cr
\bold{yi} \tab \code{numeric} \tab ADD_DESCRIPTION \cr
}  
The data frame dat.lim2014.3 contains the following columns:
\tabular{lll}{
\bold{Article} \tab \code{integer} \tab ADD_DESCRIPTION \cr
\bold{Amniotes} \tab \code{character} \tab ADD_DESCRIPTION \cr
\bold{animal} \tab \code{character} \tab ADD_DESCRIPTION \cr
\bold{N} \tab \code{integer} \tab ADD_DESCRIPTION \cr
\bold{Maternal} \tab \code{character} \tab ADD_DESCRIPTION \cr
\bold{Offspring} \tab \code{character} \tab ADD_DESCRIPTION \cr
\bold{Propagule} \tab \code{character} \tab ADD_DESCRIPTION \cr
\bold{Reproduction} \tab \code{character} \tab ADD_DESCRIPTION \cr
\bold{Environment} \tab \code{character} \tab ADD_DESCRIPTION \cr
\bold{Author} \tab \code{character} \tab ADD_DESCRIPTION \cr
\bold{Date} \tab \code{character} \tab ADD_DESCRIPTION \cr
\bold{Species} \tab \code{character} \tab ADD_DESCRIPTION \cr
\bold{Phylum} \tab \code{character} \tab ADD_DESCRIPTION \cr
\bold{Class} \tab \code{character} \tab ADD_DESCRIPTION \cr
\bold{Order} \tab \code{character} \tab ADD_DESCRIPTION \cr
\bold{Family} \tab \code{character} \tab ADD_DESCRIPTION \cr
\bold{Genus} \tab \code{character} \tab ADD_DESCRIPTION \cr
\bold{yi} \tab \code{numeric} \tab ADD_DESCRIPTION \cr
}  
The data frame dat.lim2014.4 contains the following columns:
\tabular{lll}{
\bold{Article} \tab \code{integer} \tab ADD_DESCRIPTION \cr
\bold{Amniotes} \tab \code{character} \tab ADD_DESCRIPTION \cr
\bold{animal} \tab \code{character} \tab ADD_DESCRIPTION \cr
\bold{N} \tab \code{integer} \tab ADD_DESCRIPTION \cr
\bold{Maternal} \tab \code{character} \tab ADD_DESCRIPTION \cr
\bold{Offspring} \tab \code{character} \tab ADD_DESCRIPTION \cr
\bold{Propagule} \tab \code{character} \tab ADD_DESCRIPTION \cr
\bold{Reproduction} \tab \code{character} \tab ADD_DESCRIPTION \cr
\bold{Environment} \tab \code{character} \tab ADD_DESCRIPTION \cr
\bold{Author} \tab \code{character} \tab ADD_DESCRIPTION \cr
\bold{Date} \tab \code{character} \tab ADD_DESCRIPTION \cr
\bold{Species} \tab \code{character} \tab ADD_DESCRIPTION \cr
\bold{Phylum} \tab \code{character} \tab ADD_DESCRIPTION \cr
\bold{Class} \tab \code{character} \tab ADD_DESCRIPTION \cr
\bold{Order} \tab \code{character} \tab ADD_DESCRIPTION \cr
\bold{Family} \tab \code{character} \tab ADD_DESCRIPTION \cr
\bold{Genus} \tab \code{character} \tab ADD_DESCRIPTION \cr
\bold{yi} \tab \code{numeric} \tab ADD_DESCRIPTION \cr
}  
The data dat.lim2014.phy1 contains ADD_DETAILS\cr
The data dat.lim2014.phy2 contains ADD_DETAILS\cr
The data dat.lim2014.phy3 contains ADD_DETAILS\cr
The data dat.lim2014.phy4 contains ADD_DETAILS\cr
}
\details{ADD_DETAILS}
\source{ADD_REFERENCE}
\author{ADD_CONTRIBUTOR_NAME, \email{ADD_EMAIL}}
\examples{
### copy data into 'dat' and examine data
dat <- dat.lim2014
dat
}
\keyword{datasets}
\concept{ADD_CONCEPT}
