\name{dat.lim2014}
\docType{data}
\alias{dat.lim2014.1}
\alias{dat.lim2014.2}
\alias{dat.lim2014.3}
\alias{dat.lim2014.4}
\alias{dat.lim2014.phy1}
\alias{dat.lim2014.phy2}
\alias{dat.lim2014.phy3}
\alias{dat.lim2014.phy4}
\alias{dat.lim2014}
\title{Heterogeneity in individual quality and reproductive trade‐offs within species}
\description{Correlations between: 1) maternal size and offspring size (dat.lim2014.1), 2) maternal size and maternal fecundity (dat.lim2014.2), 3) offspring size and offspring number (dat.lim2014.3), 4) offspring size and offspring number, adjusted for maternal size (dat.lim2014.4)}
\usage{dat.lim2014}
\format{
The data frame dat.lim2014.1 contains the following columns:
\tabular{lll}{
\bold{Article} \tab \code{integer} \tab a numeric identifier denoting the primary article \cr
\bold{Datapoint} \tab \code{integer} \tab a unqiue numeric identifier for the specific correlation \cr
\bold{Amniotes} \tab \code{character} \tab whether the species is amniotic or not \cr
\bold{animal} \tab \code{character} \tab the species in which the correlation is made \cr
\bold{Common.Name} \tab \code{character} \tab common name of the species \cr
\bold{N} \tab \code{integer} \tab the sample size of the correlation \cr
\bold{Maternal} \tab \code{character} \tab the trait used to quantify maternal size \cr
\bold{Offspring} \tab \code{character} \tab the trait used to quantify offspring size \cr
\bold{Reproduction} \tab \code{character} \tab does the species reproduce semalparously or iteroparously \cr
\bold{Propagule} \tab \code{character} \tab the life stage of the offspring \cr
\bold{Environment} \tab \code{character} \tab is the correlaiton quanitifed in the wild or in captivity \cr
\bold{Author} \tab \code{character} \tab authors of the primary study \cr
\bold{year} \tab \code{character} \tab year published \cr
\bold{Phylum} \tab \code{character} \tab phylum of the species \cr
\bold{Class} \tab \code{character} \tab class of the species \cr
\bold{Order} \tab \code{character} \tab order of the species \cr
\bold{Family} \tab \code{character} \tab family of the species \cr
\bold{Genus} \tab \code{character} \tab genus of the species \cr
\bold{RU} \tab \code{character} \tab single of lifetime reproductive unit \cr
\bold{yi} \tab \code{numeric} \tab effect size of correlation (zr) \cr
}  
The data frame dat.lim2014.2 contains the following columns:
\tabular{lll}{
\bold{Article} \tab \code{integer} \tab a numeric identifier denoting the primary article \cr
\bold{Datapoint} \tab \code{integer} \tab a unqiue numeric identifier for the specific correlation \cr
\bold{Amniotes} \tab \code{character} \tab whether the species is amniotic or not \cr
\bold{animal} \tab \code{character} \tab the species in which the correlation is made \cr
\bold{Common.Name} \tab \code{character} \tab common name of the species \cr
\bold{N} \tab \code{integer} \tab the sample size of the correlation \cr
\bold{Maternal} \tab \code{character} \tab the trait used to quantify maternal size \cr
\bold{Fecundity} \tab \code{character} \tab the trait used to quantify offspring numbers \cr
\bold{Reproduction} \tab \code{character} \tab does the species reproduce semalparously or iteroparously \cr
\bold{Propagule} \tab \code{character} \tab the life stage of the offspring \cr
\bold{Environment} \tab \code{character} \tab is the correlaiton quanitifed in the wild or in captivity \cr
\bold{Author} \tab \code{character} \tab authors of the primary study \cr
\bold{year} \tab \code{character} \tab year published \cr
\bold{Phylum} \tab \code{character} \tab phylum of the species \cr
\bold{Class} \tab \code{character} \tab class of the species \cr
\bold{Order} \tab \code{character} \tab order of the species \cr
\bold{Family} \tab \code{character} \tab family of the species \cr
\bold{Genus} \tab \code{character} \tab genus of the species \cr
\bold{RU} \tab \code{character} \tab single of lifetime reproductive unit \cr
\bold{yi} \tab \code{numeric} \tab effect size of correlation (zr) \cr
}  
The data frame dat.lim2014.3 contains the following columns:
\tabular{lll}{
\bold{Article} \tab \code{integer} \tab a numeric identifier denoting the primary article \cr
\bold{Amniotes} \tab \code{character} \tab whether the species is amniotic or not \cr
\bold{animal} \tab \code{character} \tab the species in which the correlation is made \cr
\bold{N} \tab \code{integer} \tab the sample size of the correlation \cr
\bold{Maternal} \tab \code{character} \tab the trait used to quantify maternal size \cr
\bold{Offspring} \tab \code{character} \tab the trait used to quantify offspring size \cr
\bold{Propagule} \tab \code{character} \tab the life stage of the offspring \cr
\bold{Reproduction} \tab \code{character} \tab does the species reproduce semalparously or iteroparously \cr
\bold{Environment} \tab \code{character} \tab is the correlaiton quanitifed in the wild or in captivity \cr
\bold{Author} \tab \code{character} \tab authors of the primary study \cr
\bold{Date} \tab \code{character} \tab year published \cr
\bold{Phylum} \tab \code{character} \tab phylum of the species \cr
\bold{Class} \tab \code{character} \tab class of the species \cr
\bold{Order} \tab \code{character} \tab order of the species \cr
\bold{Family} \tab \code{character} \tab family of the species \cr
\bold{Genus} \tab \code{character} \tab genus of the species \cr
\bold{yi} \tab \code{numeric} \tab effect size of correlation (zr) \cr
}  
The data frame dat.lim2014.4 contains the following columns:
\tabular{lll}{
\bold{Article} \tab \code{integer} \tab a numeric identifier denoting the primary article \cr
\bold{Amniotes} \tab \code{character} \tab whether the species is amniotic or not \cr
\bold{animal} \tab \code{character} \tab the species in which the correlation is made \cr
\bold{N} \tab \code{integer} \tab the sample size of the correlation \cr
\bold{Maternal} \tab \code{character} \tab the trait used to quantify maternal size \cr
\bold{Offspring} \tab \code{character} \tab the trait used to quantify offspring size \cr
\bold{Propagule} \tab \code{character} \tab the life stage of the offspring \cr
\bold{Reproduction} \tab \code{character} \tab does the species reproduce semalparously or iteroparously \cr
\bold{Environment} \tab \code{character} \tab is the correlaiton quanitifed in the wild or in captivity \cr
\bold{Author} \tab \code{character} \tab authors of the primary study \cr
\bold{Date} \tab \code{character} \tab year published \cr
\bold{Phylum} \tab \code{character} \tab phylum of the species \cr
\bold{Class} \tab \code{character} \tab class of the species \cr
\bold{Order} \tab \code{character} \tab order of the species \cr
\bold{Family} \tab \code{character} \tab family of the species \cr
\bold{Genus} \tab \code{character} \tab genus of the species \cr
\bold{yi} \tab \code{numeric} \tab effect size of correlation (zr) \cr
}  
The data dat.lim2014.phy1 contains a phylogeny for the species in dat.lim2014.1\cr
The data dat.lim2014.phy2 contains a phylogeny for the species in dat.lim2014.2\cr
The data dat.lim2014.phy3 contains a phylogeny for the species in dat.lim2014.3\cr
The data dat.lim2014.phy4 contains a phylogeny for the species in dat.lim2014.4\cr
}
\details{
	A meta-analysis on the strength of correlation between 1) maternal size and offspring size (dat.lim2014.1), 2) maternal size and maternal fecundity (dat.lim2014.2), 3) offspring size and offspring number (dat.lim2014.3), 4) offspring size and offspring number, adjusted for maternal size (dat.lim2014.4).
	
	In each case the correlation is given as the z-transformed correlation coefficient. In places these have been computed from other statsitics in the primary literature. A number of predictors related to life-history traits of the species are given. Topologies for the phylogenetic relationship between the species included in each data set are included. The data exhibit phylogenetic signals.
	}
\source{Lim, J.N., Senior, A.M. & Nakagawa, S. (2014). Heterogeneity in individual quality and reproductive trade‐offs within species. \emph{Evolution}, \bold{8}, 2306-2318.}
\author{Alistair M Senior, \email{alistair.senior@sydney.edu.au}}
\examples{
# exploring phylogeneitc effects - the metafor and ape packages are required
if((require(metafor) && require(ape))){
	
	# demonstrated with the maternal size, offpsring size correlation
	dat<-dat.lim2014.1
	phy<-dat.lim2014.phy1	
	
	# calculate the sample variance for zr
	dat$vi<-(1/sqrt(dat$N-1))^2
	
	# convert phy to ultrametric format (Grafen's method)
	phy<-compute.brlen(phy)
	
	# convert to correlation matrix
	CorMatrix <- vcv.phylo(phy, cor=TRUE)
	Cov <- CorMatrix[match(dat$animal, rownames(CorMatrix)),
	                 match(dat$animal, colnames(CorMatrix))]
	
	# ensure the random effects are factors
	dat$article <- as.factor(dat$Article)
	dat$effect.id <- as.factor(dat$Datapoint)
	dat$animal <- as.factor(dat$animal)
	
	# fit the phylogenetic multi-level meta-analysis and check the results
	PMA <- rma.mv(yi = yi , V = vi, 
	              random=list(~1|animal, ~1|article, ~1|effect.id), 
	              R = list(animal = Cov), data=dat)
	summary(PMA)
	
	# phylogenetic signals are sometimes quantified as phylogenetic heritability, or 'Pagel's lambda': sigma_phylo / sigma_total
	PMA$sigma2[1]/sum(PMA$sigma2)
	 
	 # an alternative is to include the typical sampling error in the total variance, which becomes analagous to I2 for the phylogeny
	 
}
}
\keyword{datasets}
\concept{ecology}
\concept{evolution}
\concept{phylogeny}
\concept{correlation}
