\name{dat.nakagawa2007}
\docType{data}
\alias{dat.nakagawa2007}
\title{ADD_TITLE}
\description{ADD_DESCRIPTION}
\usage{dat.nakagawa2007}
\format{
The data frame dat.nakagawa2007 contains the following columns:
\tabular{lll}{
\bold{StudyID} \tab \code{factor} \tab ADD_DESCRIPTION \cr
\bold{Place} \tab \code{factor} \tab ADD_DESCRIPTION \cr
\bold{Correlation} \tab \code{numeric} \tab ADD_DESCRIPTION \cr
\bold{SampleSize} \tab \code{integer} \tab ADD_DESCRIPTION \cr
}  
}
\details{ADD_DETAILS}
\source{ADD_REFERENCE}
\author{ADD_CONTRIBUTOR_NAME, \email{ADD_EMAIL}}
\examples{
### copy data into 'dat' and examine data
dat <- dat.nakagawa2007
dat
}
\keyword{datasets}
\concept{ADD_CONCEPT}
