\name{dat.nakagawa2015}
\docType{data}
\alias{dat.nakagawa2015.1}
\alias{dat.nakagawa2015.2}
\title{ADD_TITLE}
\description{ADD_DESCRIPTION}
\usage{dat.nakagawa2015}
\format{
The data frame dat.nakagawa2015.1 contains the following columns:
\tabular{lll}{
\bold{effect.id} \tab \code{integer} \tab ADD_DESCRIPTION \cr
\bold{species} \tab \code{character} \tab ADD_DESCRIPTION \cr
\bold{animal} \tab \code{character} \tab ADD_DESCRIPTION \cr
\bold{reference} \tab \code{character} \tab ADD_DESCRIPTION \cr
\bold{unit} \tab \code{character} \tab ADD_DESCRIPTION \cr
\bold{Trait} \tab \code{character} \tab ADD_DESCRIPTION \cr
\bold{trait_details} \tab \code{character} \tab ADD_DESCRIPTION \cr
\bold{male_mean} \tab \code{numeric} \tab ADD_DESCRIPTION \cr
\bold{male_SD} \tab \code{numeric} \tab ADD_DESCRIPTION \cr
\bold{male_n} \tab \code{integer} \tab ADD_DESCRIPTION \cr
\bold{female_mean} \tab \code{numeric} \tab ADD_DESCRIPTION \cr
\bold{female_SD} \tab \code{numeric} \tab ADD_DESCRIPTION \cr
\bold{female_n} \tab \code{integer} \tab ADD_DESCRIPTION \cr
}  
The data frame dat.nakagawa2015.2 contains the following columns:
\tabular{lll}{
\bold{effect.id} \tab \code{integer} \tab ADD_DESCRIPTION \cr
\bold{species} \tab \code{character} \tab ADD_DESCRIPTION \cr
\bold{animal} \tab \code{character} \tab ADD_DESCRIPTION \cr
\bold{reference} \tab \code{character} \tab ADD_DESCRIPTION \cr
\bold{unit} \tab \code{character} \tab ADD_DESCRIPTION \cr
\bold{Trait} \tab \code{character} \tab ADD_DESCRIPTION \cr
\bold{trait_details} \tab \code{character} \tab ADD_DESCRIPTION \cr
\bold{male_mean} \tab \code{numeric} \tab ADD_DESCRIPTION \cr
\bold{male_SD} \tab \code{numeric} \tab ADD_DESCRIPTION \cr
\bold{male_n} \tab \code{integer} \tab ADD_DESCRIPTION \cr
\bold{female_mean} \tab \code{numeric} \tab ADD_DESCRIPTION \cr
\bold{female_SD} \tab \code{numeric} \tab ADD_DESCRIPTION \cr
\bold{female_n} \tab \code{integer} \tab ADD_DESCRIPTION \cr
}  
The data frame dat.nakagawa2015.3 contains the following columns:
\tabular{lll}{
\bold{effect.id} \tab \code{character} \tab ADD_DESCRIPTION \cr
\bold{Phylum} \tab \code{character} \tab ADD_DESCRIPTION \cr
\bold{Class} \tab \code{character} \tab ADD_DESCRIPTION \cr
\bold{Family} \tab \code{character} \tab ADD_DESCRIPTION \cr
\bold{Host.species} \tab \code{character} \tab ADD_DESCRIPTION \cr
\bold{Host.taxon} \tab \code{character} \tab ADD_DESCRIPTION \cr
\bold{CMean} \tab \code{numeric} \tab ADD_DESCRIPTION \cr
\bold{CSD} \tab \code{numeric} \tab ADD_DESCRIPTION \cr
\bold{CN} \tab \code{integer} \tab ADD_DESCRIPTION \cr
\bold{EMean} \tab \code{numeric} \tab ADD_DESCRIPTION \cr
\bold{ESD} \tab \code{numeric} \tab ADD_DESCRIPTION \cr
\bold{EN} \tab \code{integer} \tab ADD_DESCRIPTION \cr
\bold{Transmission.by.predation.vector.or.not} \tab \code{character} \tab ADD_DESCRIPTION \cr
\bold{spontaneous.or.stimulus.response} \tab \code{character} \tab ADD_DESCRIPTION \cr
\bold{Behaviour.details} \tab \code{character} \tab ADD_DESCRIPTION \cr
\bold{Year.of.publication} \tab \code{integer} \tab ADD_DESCRIPTION \cr
\bold{Reference} \tab \code{character} \tab ADD_DESCRIPTION \cr
\bold{Expected.direction.of.change} \tab \code{character} \tab ADD_DESCRIPTION \cr
}  
The data dat.nakagawa2015.phy1 contains ADD_DETAILS\cr
The data dat.nakagawa2015.phy2 contains ADD_DETAILS\cr
}
\details{ADD_DETAILS}
\source{ADD_REFERENCE}
\author{ADD_CONTRIBUTOR_NAME, \email{ADD_EMAIL}}
\examples{
### copy data into 'dat' and examine data
dat <- dat.nakagawa2015
dat
}
\keyword{datasets}
\concept{ADD_CONCEPT}
